Split out app from runner for testing purposes
